Update behat tests
